+++ src/game/game_objects/BrickWall.cpp
@@ -78,7 +78,7 @@ void BrickWall::render() const
 	renderBrick(EBrickLocation::BottomRight);
 }
 
-void BrickWall::update(const uint64_t delta)
+void BrickWall::update(const double delta)
 {
 }
