+++ src/game/game_objects/GameObject.h
@@ -5,7 +5,7 @@ class GameObject
 public:
 	GameObject(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
 	virtual void render() const = 0;
-	virtual void update(const uint64_t delta) = 0;
+	virtual void update(const double delta) = 0;
 	virtual ~GameObject();
 protected:
 	glm::vec2 m_position;
